Fix for GitHub Issue #778:
1. Fix Garbage Collector to NOT hold its gclock the entire time that it runs. Instead, only hold it for very brief periods when absolutely necessary, and to then release it as soon as possible.

2. The ultimate fix however, was to introduce a new combined Garbage Collector and Dasd Hardener lock that each acquires and holds for the entire time that each of them runs, so as to prevent either of them from running at the same time.

I did not bother to figure out where (and with what) the interference between the two was occurring, but simply decided to use the "hammer approach" and prevent them both from running at the same time  instead. Besides, neither of them runs for very long. Each usually only runs for less then a second, so any delay in each of their "passes" (cycles) is VERY brief and hardly noticeable.